+++ dexterous_developer_cli/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod cross;
 pub mod existing;
-pub mod generate_temporary_lib;
 pub mod paths;
 pub mod remote;
 pub mod serve;
+pub mod temporary_manifest;