+++ dexterous_developer_example/src/lib.rs
@@ -82,7 +82,7 @@ fn setup_cube(
     let cube_color = Color::ORANGE;
 
     #[cfg(not(feature = "orange"))]
-    let cube_color = Color::RED;
+    let cube_color = Color::PINK;
 
     // cube
     commands.spawn((